Added maven local publishing.
