+++ .idea/artifacts/lib_jvm_1_1_0_beta01.xml
@@ -1,0 +1,8 @@
+<component name="ArtifactManager">
+  <artifact type="jar" name="lib-jvm-1.1.0-beta01">
+    <output-path>$PROJECT_DIR$/lib/build/libs</output-path>
+    <root id="archive" name="lib-jvm-1.1.0-beta01.jar">
+      <element id="module-output" name="IdeaProjects.apiclient.lib.jvmMain" />
+    </root>
+  </artifact>
+</component>
