+++ core/gradle.properties
