--- lib/build.gradle.kts
@@ -1,46 +1,0 @@
-plugins {
-    alias(libs.plugins.kotlin.multiplatform)
-    alias(libs.plugins.kotlinx.serialization)
-//    alias(libs.plugins.vaniktek.mavem.publish)
-    alias(libs.plugins.jetbrains.dokka)
-    alias(libs.plugins.android.library)
-}
-
-kotlin {
-    androidTarget()
-    iosX64()
-    iosArm64()
-    iosSimulatorArm64()
-    jvm()
-
-    sourceSets {
-        commonMain {
-            dependencies {
-                implementation(libs.kotlinx.coroutines.core)
-                implementation(libs.mready.json)
-                implementation(libs.kotlinx.serialization.json)
-
-                //Ktor
-                implementation(libs.ktor.client.core)
-                implementation(libs.ktor.client.cio)
-                implementation(libs.ktor.client.contentNegotiation)
-                implementation(libs.ktor.serialization.json)
-            }
-        }
-        commonTest {
-            dependencies {
-                implementation(kotlin("test"))
-            }
-        }
-    }
-}
-
-android {
-    namespace = "net.mready.apiclient"
-    compileSdk = 34
-
-    compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_17
-        targetCompatibility = JavaVersion.VERSION_17
-    }
-}
