--- lib/src/commonMain/kotlin/client/errors.kt
@@ -1,11 +1,0 @@
-package client
-
-/**
- * Exception thrown when the server returns an error code.
- */
-class HttpCodeException(val code: Int, message: String) : RuntimeException(message)
-
-/**
- * Exception thrown when the response body cannot be parsed.
- */
-class ParseException(message: String, cause: Throwable?) : RuntimeException(message, cause)
