--- lib/src/iosTest/kotlin/NSDataTest.kt
@@ -1,16 +1,0 @@
-import io.ktor.utils.io.core.*
-import kotlin.test.Test
-import kotlin.test.assertEquals
-
-
-class NSDataTest {
-
-    @Test
-    fun testNsDataConversions() {
-        val helloString = "Hello, World!"
-        val byteArray = helloString.toByteArray()
-
-        assertEquals(helloString, byteArray.decodeToString())
-        assertEquals(helloString, byteArray.toNSData().toByteArray().decodeToString())
-    }
-}
